+++ imdb/pascal_voc.py
@@ -1,7 +1,7 @@
 import os
 import numpy as np
 
-from symnet.logger import logger
+from net.logger import logger
 from .imdb import IMDB
